Sanitary Napkin Manufacturing Company in India
The World Health Organization defines a female as an 
adolescent if she is between the ages of 10 and 19. The 
time between childhood and adulthood known as 
adolescence is marked by the child's growth and 
development. During this time, the child's physical, 
psychological, and biological growth happens. It is 
understood to be a special moment in a girl's life that 
calls for extra consideration. Menarche, which marks 
the beginning of a woman's reproductive phase, is an 
important biological milestone in her life. Menarche 
often occurs between the ages of 12 and 13, and this 
age is highly consistent across populations. 
Unfortunately, many women are unable to control their 
periods due to a lack of knowledge about menstruation 
preparation and management or a lack of resolve.
Menstrual wastes are the wastes a female produces 
throughout her reproductive years. These wastes are 
produced during menstruation, often known as menses, 
periods, or the monthly bleeding cycle. The three 
phases of the menstrual cycle are the follicular phase 
(proliferative), ovulation phase, and lacteal phase 
(secretary). Menstruation is regulated by hormones; 
during this time, the uterine lining, or endometrium, 
thickens and sheds, causing bleeding that typically lasts 
3-5 days but can occasionally last up to 7 days. During 
menstruation, the endometrial lining is destroyed in 
two-thirds. Along with blood, menstrual fluid also 
contains mucus and vaginal secretions.
Every woman's menstrual flow is different; it may be 
heavier or lighter at the beginning of the cycle or 
fluctuate during the cycle. Menstrual blood can be 
crimson, vivid red, dark brown, or even black. 
Menstrual blood may or may not smell bad when it 
comes in contact with air. Before and after menopause, 
as well as during gynaecological cancers, menstrual flow 
and duration change. In the presence of fibroids, 
polyps, and endometriosis, menstrual flow rises, and 
excessive blood loss during menstruation can cause 
anaemia.
Women have developed their own unique strategies to 
deal with this era. These approaches are influenced by a 
variety of factors, including individual preferences, the 
accessibility of resources, economic status, cultural 
beliefs and practises, level of education, and knowledge 
about menstruation. Inadequate menstrual hygiene 
practises can result in toxic shock syndrome, 
reproductive tract infections (RTI), and other vaginal 
problems, which are serious health risks Poor genital 
hygiene has a substantial impact on adolescents' health. 
Most girls are unprepared for menarche because they 
are misled or ignorant about menstruation.
This study's main objective was to provide an overview 
of the problems and potential solutions for handling 
menstrual waste in low-income countries. The review 
article's objective was to gain more knowledge 
regarding menstrual behaviours, product design, 
consumer demand, and disposal techniques. Along with 
an analysis of current knowledge in public health, water 
and sanitation, and solid waste management, it includes 
a description of modern menstrual hygiene needs and 
management.